refactor(emir): design 链审计收敛——业务层裸指针清退 + id 强类型尾巴（B 组残留）
design 链审计（docs/emir/raw-type-pointer-audit-2026-09-15.md，报告
一并入册）B 组残留逐项收敛：

- B-5a ds_name_mapper 树观察裸指针清退：DSNameMapperT 成员/构造/
  set_tree 改 CMSharedPtr<const DSHierTree>；ds_make_name_mapper 收
  CMSharedPtr<const DSDesign>，树观察 = aliasing shared_ptr 持 design
  计数（29cbe0f 先例形态）——mapper 生命周期自持，调用侧不再背
  「design 存活期覆盖 mapper」契约（Python 面同步受益）
- B-5b set_block_hasher 主口键与 injected_ 表键 uint32_t → CMCellId
  （hasher 底座裸值域豁免边界 CMCellId{} 显式互转）
- B-5c DSNameMapperT 双实例化拆分：**评估后不做**——IdT 同时是 hasher
  底座 DSNameHasherT<IdT> 的实例化参数，单拆 mapper 需 hasher 底座增
  CMInstanceId/CMNetId 实例化（B-4 主战场牵动，超本批范围），留待
  hasher 强类型化一并裁定
- B-6 ds_id_map 消混流：DSIdPartitionSliceT<IdT> 模板双实例化
  （DSInstIdPartitionSlice/DSNetIdPartitionSlice 分表，id 强类型）；
  ds_collect_partition_id_slice 的 bool instance_kind 参数随分型拆除
  （ds_collect_inst_id_slice / ds_collect_net_id_slice——参数面各取
  所需）；merge 模板显式实例化两维度；export/Python 面拆双桥
- B-7 DSDensityWeights::layer_factors_ 键 CMLayerId::int_type →
  CMLayerId（CMUnorderedMap）
- B-9 ds_verify_partition 签名 uint32_t partition_id → CMPartitionId
- B-10 两 pipeline ctx 环境裸指针清退（design/stack/block_data/
  net_data/fake_cells → CMSharedPtr 共享注入；ds_parse_def_components/
  ds_parse_def_nets 签名随迁）；orient 改 GEOOrientation 直传
  （ds_transform_util place_from_def 参数枚举化——typed→int→typed
  往返消除，整型转枚举收拢在 Si2 回调边界一次完成）
- B-12 借用指针集 8 处清退：CMVector<const T*>& 全部改
  CMVector<CMSharedPtr<const T>>（ds_build_hier_tree/
  ds_merge_global_density/ds_collect_net_union_slice/ds_build_net_union/
  finalize_from_flatten/ds_merge_id_partition_slices/ds_verify_design/
  ds_make_name_mapper）——export 桥改 shared 收集（nanobind shared_ptr
  caster，调用侧本就持 shared）
- S7 哨兵防御：ds_collect_net_union_slice 聚合循环跳过无有效 block
  cell id 的树节点（root 占位防误命中无效键）
- 杂项：get_full_name(0) Python 面 None 形态补注（design-knowledge.md）；
  7 处「全局平铺 pin id」旧口径注释统一 2026-09-16 裁定 3 口径
  （ds_name_hasher.h/ds_types.h/ds_flatten.h/design_export.cpp）

修复在途崩溃：ds_parse_def_components export 桥对 move 后空 shared_ptr
解引用（QA 实测 SIGSEGV，注入侧改拷贝 shared 保留出参句柄）。

序列化字节影响：id 换型全部走 .value()/CMXxxId{} 直通同宽（StrongIdT
手写直通），正式对象落盘布局零变化；DSIdPartitionSliceT 为临时对象
（S9 单次 flow 内写读配对）。测试侧新增 tests/borrow_view.h（栈对象
aliasing 借用观察辅助，BUILD 逐 target 登记）。

+++ src/emir/design/cpp/ds_instance_pipeline.cpp
@@ -106,9 +106,9 @@ void DSInstanceBuildNode::handle(DSInstanceContext& ctx) {
     }
     DSInstance inst;
     inst.set_cell_id(ctx.cell_id);
-    // R6：t → pos 放置边界换算（orient 直转已在适配层完成）
-    inst.set_transform(place_from_def(ctx.placement,
-                                      static_cast<int>(ctx.orient),
+    // R6：t → pos 放置边界换算（GEOOrientation 直传——评审 B-10，消
+    // typed→int→typed 往返）
+    inst.set_transform(place_from_def(ctx.placement, ctx.orient,
                                       ctx.cell_bbox, ctx.cell_origin_x,
                                       ctx.cell_origin_y));
     inst.set_placement_status(ctx.placement_status);
